Made the  periodic table different
I don’t know if you will like this so I just uncommented the old code
if you want to change back. I think we should look in to putting the
info into the space above the table or adding a scroll bar. I also
think the electron shell configuration should be put in as part of that
info.